Fix build with clang
If we don't cast to unsigned int clang complains with following message:

error: case value evaluates to 2690980318, which cannot be narrowed to
type 'int'

+++ src/servicemodel.cpp
@@ -87,7 +87,7 @@ QVariant ServiceModel::data(const QModelIndex &index, int role) const
         return QVariant();
     }
     const QList<RemoteService::Ptr> srv = d->m_browser->services();
-    switch (role) {
+    switch ((uint)role) {
     case Qt::DisplayRole:
         switch (index.column()) {
         case ServiceName: return srv[index.row()]->serviceName();
